fix: change sensorValueRaw to float in Lift to prevent integer truncation of ADC voltage

+++ lib/Lift/Lift.h
@@ -62,7 +62,7 @@ public:
 
 private:
     float sensorValue;
-    int sensorValueRaw;
+    float sensorValueRaw;
     char upPin;
     char downPin;
     uint8_t dacAddress;
